% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_scores-data.R
\docType{data}
\name{box_scores}
\alias{box_scores}
\title{descr}
\format{
An object of class `r class(box_scores)`, with `r ncol(box_scores)` columns and `r nrow(box_scores)` rows.

\tabular{ll}{
   \code{site} \tab Name of the site\cr
   \tab \cr
   \code{sp} \tab Name of the "species" scored. It can represent a taxon (species or genus) or functional group. \cr
   \tab \cr
   \code{status} \tab numerical EBQI score assigned to "sp". 0 is bad and 4 is good. \cr
}
}
\usage{
data(box_scores)
}
\description{
moer exple
}
\examples{
data(box_scores)

head(box_scores)
summary(box_scores)
str(box_scores)

}
\references{
paper
}
\keyword{datasets}
