% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-v0_2.R
\name{classify}
\alias{classify}
\title{Ecological status class for EBQI}
\usage{
classify(
  x,
  labs = c("bad", "poor", "moderate", "good", "high"),
  brks = c(0, 3.5, 4.5, 6, 7.5, 11),
  use_lab = TRUE
)
}
\arguments{
\item{x}{vector storing the EBQI values to be converted in ecological status class. Can be of lenght of 1 or longer.}

\item{labs}{character vector with labels for ecological status classes. Default to c('bad','poor','moderate','good','high')}

\item{brks}{numeric vevtor with breaks to bin EBQI numerical values. Default to c(0,3.5,4.5,6,7.5,11)}

\item{use_lab}{logical. set "use_lab" equal TRUE to use labels as in `labs`. use_lab=FALSE it returns a numeric value (1= bad to 5 =high)}
}
\value{

}
\description{
function to classify the EBQI into ecological status.
}
\examples{
EIs<-c(1,0,9.5,3.5,7,5)

classify(EIs,use_lab=TRUE)
classify(EIs,use_lab=FALSE)

}
\references{
Personnic, S., Boudouresque, C.F., Astruch, P., Ballesteros, E., Blouet, S., Bellan-Santini, D., Bonhomme, P., Thibault-Botha, D., Feunteun, E., Harmelin-Vivien, M., Pergent, G., Pergent-Martini, C., Pastor, J., Poggiale, J.-C., Renaud, F., Thibaut, T., Ruitton, S., 2014. An Ecosystem-Based Approach to Assess the Status of a Mediterranean Ecosystem, the Posidonia oceanica Seagrass Meadow. PLoS One 9, e98994. https://doi.org/10.1371/journal.pone.0098994
}
\seealso{
\code{\link{EBQI}}
}
\author{
Filippo Ferrario, \email{filippo.ferrario.1@ulaval.ca}
}
