% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBQI-v1_2.R
\name{EBQI}
\alias{EBQI}
\title{Calculate the Ecosystem-Based Quality Index}
\usage{
EBQI(
  score_dataset,
  weights_dataset,
  model_description,
  SITE = NULL,
  DATA_SP = NULL,
  SCORE = NULL,
  W_BOX = NULL,
  WEIGHTS = NULL,
  MAX_weight = 10,
  MOD_BOX = NULL,
  MODEL_SP = NULL,
  PERTURBATION = 0
)
}
\arguments{
\item{score_dataset}{dataframe with the scores of each box in the EBQI model. `score_dataset` must include the following columns: site name, species name, score value for each species.}

\item{weights_dataset}{dataframe with the weights of each box in the EBQI in the model. Different models and weighting scenarios can be specified. `weights_dataset` must include the following columns: model name, boxes' names, weighting scenario and weight value.}

\item{model_description}{dataframe specifying which species are considered in each box of the EBQI model. Different models can be specified.`model_description` must include the following columns: model name, boxes' names and species names.}

\item{SITE}{character. Name of the column in `score_dataset` storing the name of the site for which to calculate the EBQI.}

\item{DATA_SP}{character. Name of the column in`score_dataset` storing the names of the species.}

\item{SCORE}{character. Name of the column in `score_dataset` storing the species' scores.}

\item{W_BOX}{character. Name of the column in `weights_dataset` storing the box names.}

\item{WEIGHTS}{character. Name of the column in `weights_dataset` storing the numerical weights.}

\item{MAX_weight}{numeric. Maximum value of the weights. So far it is assumed that the weight values for the different model*scenario have a common range.}

\item{MOD_BOX}{character. Name of the column in `model_description` storing the box names.}

\item{MODEL_SP}{character. Name of the column in`model_description` storing the names of the species.}

\item{PERTURBATION}{numerical value to be added to weighs to introduce perturbations. Either one number or a vector of lenght equal to the number of boxes considered in the model in use.}
}
\value{
A named list of two elements. `.$ebqi` stores the numerical value of the EBQI index. `.$scores` strores the (mean) scores used to calculate the EQBI.
}
\description{
Calculate the Ecosystem-Based Quality Index using the formula given in Personnic et al (2014).
}
\details{
The three input datasets `score_dataset`,`weights_dataset`,`model_description` can be used to calculate EBQI based on different model and weighting scheme specifications. A model is specified by the boxes considered. Different weighting scenario can be specified for the same model.
The names of the model boxes must match between input dataframes.
Only boxes for which there are species scored in the `score_dataset` will be actually used. 
When 2 or more species are included in the same box their scores are averaged.

For a real case study using this package please see \url{https://dataverse. } and the related paper.
}
\examples{

# create input datasets
scores<-data.frame(site=rep('s1',9),
                   sp=c('MPO','arenicola_marina','asteriidae','cancer_irroratus','echinarachnius_parma','strongylocentrotus_droebachiensis','OT-Invertivorous_Invertebrates','Suspension_feeders','Bioturbation'),
                   parameter=rep('density',9), unit='ind/m-2',
                   status=c(2,0,2,4,2,4,4,0,3))

weights<-data.frame(model='m1', scenario='scen1',box=c('MPO','Herbivore','Detritus feeder','Invertivorous Invertebrates','Infauna','OT-Invertivorous Invertebrates','Bioturbation','Suspension feeder'), weight_1_10=c(10,7,3,6,4,1,1,1))

mods<- data.frame(model='m1',
                     species=c('MPO','arenicola_marina','asteriidae','cancer_irroratus','echinarachnius_parma','strongylocentrotus_droebachiensis','OT-Invertivorous_Invertebrates','Suspension_feeders','Bioturbation','Nereis_sp'),
                     boxes=c('MPO','Detritus feeder','Invertivorous Invertebrates','Invertivorous Invertebrates','Detritus feeder','Herbivore','OT-Invertivorous Invertebrates','Suspension feeder','Bioturbation','Infauna'),
                     box_ID=NA) 
# run the function
EBQI(score_dataset=  scores ,weights_dataset=weights, model_description= mods,
      SITE='site',DATA_SP='sp',SCORE='status',
      W_BOX='box' ,WEIGHTS='weight_1_10' ,
      MOD_BOX='boxes',MODEL_SP='species')


}
\references{
Personnic, S., Boudouresque, C.F., Astruch, P., Ballesteros, E., Blouet, S., Bellan-Santini, D., Bonhomme, P., Thibault-Botha, D., Feunteun, E., Harmelin-Vivien, M., Pergent, G., Pergent-Martini, C., Pastor, J., Poggiale, J.-C., Renaud, F., Thibaut, T., Ruitton, S., 2014. An Ecosystem-Based Approach to Assess the Status of a Mediterranean Ecosystem, the Posidonia oceanica Seagrass Meadow. PLoS One 9, e98994. https://doi.org/10.1371/journal.pone.0098994
}
\seealso{
\code{\link{classify}}
}
\author{
Filippo Ferrario, \email{filippo.ferrario.1@ulaval.ca}
}
