% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EzBQI_package.R
\docType{package}
\name{EzBQI-package}
\alias{EzBQI}
\alias{EzBQI-package}
\title{The EzBQI package}
\description{
Functions to easily calculate the Ecosystem-Based Quality Index and obtain the ecological status class.
}
\references{

}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
\keyword{internal}
