% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods-data.R
\docType{data}
\name{mods}
\alias{mods}
\title{descr}
\format{
An object of class `r class(mods)`, with `r ncol(mods)` columns and `r nrow(mods)` rows.


\tabular{ll}{
   \code{sp} \tab name of the taxonomical or functional group in an EBQI model\'s box\cr
   \tab \cr
   \code{ID_box} \tab numeric identifer of the EBQI model\'s box \cr
   \tab \cr
   \code{box} \tab name of the EBQI model\'s box \cr
   \tab \cr
   \code{model} \tab name of the EBQI ecosystem model\cr
}
}
\usage{
data(mods)
}
\description{
moer exple
}
\examples{
data(mods)

head(mods)
summary(mods)
str(mods)

}
\references{
paper
}
\keyword{datasets}
