% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights-data.R
\docType{data}
\name{weights}
\alias{weights}
\title{descr}
\format{

}
\usage{
data(weights)
}
\description{
moer exple
}
\examples{
data(weights)

head(weights)
summary(weights)
str(weights)

}
\references{
paper
}
\keyword{datasets}
